Change S3 concern into a lib, for use in controllers too
